Refactor some commands into a more native kotlin way, new command api, some changes to the old api for compatibility.

+++ core/src/main/kotlin/me/viniciusarnhold/altaria/utils/Timers.kt
@@ -1,18 +1,14 @@
 package me.viniciusarnhold.altaria.utils
 
-import commands.MessageUtils
 import org.apache.logging.log4j.LogManager
 import sx.blah.discord.handle.obj.IMessage
-import sx.blah.discord.util.DiscordException
-import sx.blah.discord.util.MissingPermissionsException
-import sx.blah.discord.util.RateLimitException
 import java.util.*
 import java.util.concurrent.TimeUnit
 
 /**
  * Created by Vinicius.
 
- * @since ${PROJECT_VERSION}
+ * @since 1.0
  */
 object Timers {
 
@@ -24,7 +20,7 @@ object Timers {
             schedule(message, TimeUnit.MILLISECONDS.convert(time, originUnit))
         }
 
-        @JvmOverloads fun schedule(message: IMessage, nanos: Long = 60000L) {
+        fun schedule(message: IMessage, nanos: Long = 60000L) {
             timer.schedule(newMessageDeletionTask(message), nanos)
         }
 
@@ -35,44 +31,13 @@ object Timers {
         private fun newMessageDeletionTask(message: IMessage): TimerTask {
             return object : TimerTask() {
                 override fun run() {
-                    try {
+                    if (!message.isDeleted) {
                         message.delete()
-                    } catch (e: MissingPermissionsException) {
-                        MessageUtils.handleDiscord4JException(logger, e)
-                    } catch (e: RateLimitException) {
-                        MessageUtils.handleDiscord4JException(logger, e)
-                    } catch (e: DiscordException) {
-                        MessageUtils.handleDiscord4JException(logger, e)
                     }
-
                 }
             }
         }
 
-        val logger = LogManager.getLogger()!!
-    }
-
-    object CacheCleanUpService {
-
-        private val cacheCleanUpTimer: Timer = Timer("Cache cleanup Timer Service.", true)
-
-        fun <K, V> schedule(mapToClean: MutableMap<K, V>, keyToClean: K, expectedValue: V) {
-            schedule(mapToClean, keyToClean, expectedValue, 1, TimeUnit.HOURS)
-        }
-
-        fun <K, V> schedule(mapToClean: MutableMap<K, V>, keyToClean: K, expectedValue: V, time: Long, unit: TimeUnit) {
-            cacheCleanUpTimer.schedule(
-                    createNewCleanUpTask(mapToClean, keyToClean, expectedValue),
-                    TimeUnit.NANOSECONDS.convert(time, unit)
-            )
-        }
-
-        private fun <K, V> createNewCleanUpTask(mapToClean: MutableMap<K, V>, keyToClean: K, expectedValue: V): TimerTask {
-            return object : TimerTask() {
-                override fun run() {
-                    mapToClean.remove(keyToClean, expectedValue)
-                }
-            }
-        }
+        val logger = LogManager.getLogger()
     }
 }
